// src/app/modules/openapi/model/paged-response-user-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDTO } from './user-dto';


export interface PagedResponseUserDTO { 
    content?: Array<UserDTO>;
    currentPage?: number;
    totalPages?: number;
    totalElements?: number;
    pageSize?: number;
    first?: boolean;
    last?: boolean;
    empty?: boolean;
    sortField?: string;
    sortDir?: string;
    keyword?: string;
    searchParams?: { [key: string]: any; };
    allowedSortFields?: Set<string>;
}